AFAIK classes are fully intergrated now, ready to merge.
also fixed annoying on hover dissapear CSS thing i had from a while back.
