+++ app/index.html
@@ -34,7 +34,13 @@
         </div>
     </div>
 
-    <!-- Javascript -->
+    <!--A poor man's module.exports -->
+    <script src='./js/init.js'></script>
+    <script src='./js/melodiiDir.js'></script> 
+    <script src='./js/melodiiButtons.js'></script>
+    <script src='./js/melodii.js'></script>
+    <script src='./js/melodiiCNTRL.js'></script>
+    <script src='./js/melodiiDOM.js'></script>
     <script src='./js/main.js'></script>
 </body>
